Add email-smtp feature and use tera templates for mail
This feature allows to send email via smtp with the help of the lettre crate.

In the process some environment variable are expected to be set.

A new trait `Configurable` with the `check_conf` method allow to verify the
configuration upfront when starting the server in order to prevent the misconfiguration
to be detected on production.

The mail templates need both HTML and text temlpate in order to be able to send
accessible mails.

Also both web and email templates now have all environment variables starting
with `JELLY_` in their context when being rendered.